minor changes
camera ; default constractor and for a pointer to camera .
scene ; add load_cam that push a camera to the back, and a default constractor, and checking in creating for renderer if there is no camera at all for the scene and create a new camera.
changed position for some files of grafical  objects.
btw there is a resizing and reposition of the object that is from the drawTringals in the renderer