I fixed the transformation and the memory alocation
there was some problam in the create transformation that I hoped I fixed.
fixed some memory allocation problem (added delete of some pointers in  Renderer::DrawTriangles
so now there isn't an increase in the memory being used when the code rans).
added some functions to accese the Transformations in camera (we can write all the transformations in a different class ).