added shading options
there are flate shading,
Gouraud Shading and
Phong Shading.
there are some problames with some normals so they don't look as good as they should (some normals that we get from the models (get vertex norm and
get face norm ) are zero , or aprroching zero , or just nan (not a number )).
but still it looks surprisingly fine on the moods (exept flat).
we need to add specular ligth to all the moods .
we probably need to add more properties to the ligth class to allow transformations for the ligth .
we need to add z buffer , and some more stuff by the requairment.