fixed  flat shading + phong shading
Fixed the flate shading , We just switch in the midel of the for of if (type==0) , between some valuse . it was surprisingly not so noticebel  when running the code.
also I wrote the code for the phong shading  , it was almost identical to the code for chosing the color for flat shading (in the put pixel 3 , and that was the only thing I added there, exept passing valus for it to run proparly ).
there were only some minor changes that happend from the pact that we passed different parameters .
the phong looks extremly good .